lwip: FIONREAD for UDP sockets (int value + LINUXMODE default) — Quake multiplayer
Two coupled fixes so ioctl(FIONREAD) works on Phoenix UDP sockets:
- sockets.h: FIONREAD = _IOR('f',127,int) (was unsigned long). _IOR encodes sizeof(type) into
  the command value, so unsigned-long(8) made it a DIFFERENT number than libphoenix
  <sys/ioctl.h>'s int(4) that clients send -> lwip_ioctl's `case FIONREAD` never matched ->
  ENOSYS. FIONBIO already matched (both ulong), which is why non-blocking worked but FIONREAD
  didn't. lwip_ioctl writes *((int*)argp), so int is correct.
- opt.h: LWIP_FIONREAD_LINUXMODE default 0->1. lib-lwip core (lwip-core.a) is compiled without
  the port's -Iinclude/default-opts, so the port lwipopts.h define does NOT reach lwip_ioctl();
  the opt.h default does. Enables FIONREAD for UDP (next datagram size, buffered into lastdata).

Without these, Quake's UDP server hung at "Loading" with "UDP: ioctlsocket (FIONREAD) failed
(ENOSYS)". HW-validated: server reaches protocol-666 init, no ENOSYS, 0 faults.

+++ src/include/lwip/sockets.h
@@ -418,7 +418,12 @@ typedef struct ipv6_mreq {
 #endif /* !defined(FIONREAD) || !defined(FIONBIO) */
 
 #ifndef FIONREAD
-#define FIONREAD    _IOR('f', 127, unsigned long) /* get # bytes to read */
+/* Phoenix: int (not unsigned long) to match libphoenix <sys/ioctl.h> FIONREAD=_IOR('f',127,int).
+ * _IOR encodes sizeof(type) into the command value, so ulong(8) vs int(4) made the client's
+ * FIONREAD a DIFFERENT number than lwip_ioctl's `case FIONREAD` -> no match -> ENOSYS (Quake UDP
+ * server "ioctlsocket (FIONREAD) failed (ENOSYS)" hang at Loading). FIONBIO already matched (both
+ * ulong), which is why non-blocking worked but FIONREAD didn't. lwip_ioctl writes *((int*)argp). */
+#define FIONREAD    _IOR('f', 127, int) /* get # bytes to read */
 #endif
 #ifndef FIONBIO
 #define FIONBIO     _IOW('f', 126, unsigned long) /* set/clear non-blocking i/o */
